Played around with slopedepthbias

+++ DXR-Project/D3D12/D3D12GraphicsPipelineState.h
@@ -15,7 +15,7 @@ struct GraphicsPipelineStateProperties
 	Uint32						NumRenderTargets	= 0;
 	DXGI_FORMAT					DepthBufferFormat	= DXGI_FORMAT_UNKNOWN;
 	bool						EnableDepth			= false;
-	Float32						DepthBias			= D3D12_DEFAULT_DEPTH_BIAS;
+	Int32						DepthBias			= D3D12_DEFAULT_DEPTH_BIAS;
 	Float32						SlopeScaleDepthBias	= D3D12_DEFAULT_SLOPE_SCALED_DEPTH_BIAS;
 	Float32						DepthBiasClamp		= D3D12_DEFAULT_DEPTH_BIAS_CLAMP;
 	bool						EnableBlending		= false;
